server/cmd/server/root_test.go
package server

import (
	"github.com/spf13/cobra"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestRootCmdExecute(t *testing.T) {
	// Mock version and exit function
	//mockVersion := version.Version{Major: 1, Minor: 0, Patch: 0}
	//exitCalled := false
	//mockExit := func(code int) {
	//	exitCalled = true
	//}
	//
	//// Create a new rootCmd instance
	//root := newRootCmd(mockVersion, mockExit)
	//
	//// Execute rootCmd with arguments
	//args := []string{"argument1", "argument2"}
	//root.Execute(args)
	//
	//// Validate that the exit function was not called
	//assert.False(t, exitCalled, "Exit should not have been called")

	// TODO: Add more assertions based on the behavior you want to test.
	// For example, you can use assertions from the testify/assert package to check the expected behavior.
	// assert.Equal(t, expectedValue, actualValue, "Description of the assertion")

	// Note: The provided code does not have a lot of complex behavior,
	// so you may need to extend the test based on your application's requirements.
}

func TestShouldPrependRun(t *testing.T) {
	// Mock cobra.Command
	mockCmd := &cobra.Command{
		Use: "mock-cmd",
	}

	// Test case: should not prepend "run" when help or version flags are present
	assert.False(t, shouldPrependRun(mockCmd, []string{"-h"}), "Should not prepend run for help flag")
	assert.False(t, shouldPrependRun(mockCmd, []string{"--help"}), "Should not prepend run for help flag")
	assert.False(t, shouldPrependRun(mockCmd, []string{"-v"}), "Should not prepend run for version flag")
	assert.False(t, shouldPrependRun(mockCmd, []string{"--version"}), "Should not prepend run for version flag")

	// Test case: should prepend "run" for other arguments
	assert.True(t, shouldPrependRun(mockCmd, []string{"arg1", "arg2"}), "Should prepend run for regular arguments")
}
